agregando crud de roles y permisos en laravel

+++ iijp-api/app/Http/Controllers/Api/Admin/UserController.php
@@ -11,9 +11,21 @@ class UserController extends Controller
 
     public function index()
     {
-        $data = User::all();
-        return response()->json($data, 200);
+
+        $users = User::whereDoesntHave('roles', function ($query) {
+            $query->where('name', 'admin');
+        })->paginate(20);
+
+
+        $users->getCollection()->transform(function ($user) {
+            $user->role = $user->getRoleNames()->first();
+            return $user;
+        });
+
+        return response()->json($users, 200);
     }
+
+
 
 
     public function show($id)
@@ -25,35 +37,74 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|email|unique:users,email',
             'password' => 'required|min:8',
+            'role' => 'required|string|max:255|exists:roles,name',
         ]);
 
-        $validatedData['password'] = bcrypt($validatedData['password']);
-        $user = User::create($validatedData);
-        $user->assignRole('admin');
 
-        return response()->json($user, 201);
+        $validatedData['password'] = bcrypt($validatedData['password']);
+
+        $user = User::create($validatedData);
+
+        $user->assignRole($validatedData['role']);
+
+        return response()->json($user->only(['id', 'name', 'email', 'role']), 201);
     }
 
 
     public function update(Request $request, $id)
     {
+
+        $validatedData = $request->validate([
+            'name' => 'nullable|string|max:255',
+            'email' => 'nullable|email|unique:users,email,' . $id,
+            'password' => 'nullable|min:8',
+            'role' => 'nullable|string|max:255|exists:roles,name',
+        ]);
+
         $user = User::findOrFail($id);
-        $user->fill($request->all());
+
+        if ($request->has('name')) {
+            $user->name = $validatedData['name'];
+        }
+
+        if ($request->has('email')) {
+            $user->email = $validatedData['email'];
+        }
+
+        if ($request->has('password')) {
+            $user->password = bcrypt($validatedData['password']);
+        }
+
+        if ($request->has('role')) {
+            $user->syncRoles([$validatedData['role']]);
+        }
+
         $user->save();
 
-        return response()->json($user, 200);
+        return response()->json($user->only(['id', 'name', 'email', 'role']), 200);
     }
 
 
     public function destroy($id)
     {
         $user = User::findOrFail($id);
+
+        if ($user->hasRole('admin')) {
+            return response()->json([
+                'message' => 'No se puede eliminar un usuario administrador.',
+            ], 403);
+        }
+
         $user->delete();
 
-        return response()->json(['message' => 'Usuario eliminado correctamente'], 200);
+        return response()->json([
+            'message' => 'Usuario eliminado correctamente',
+        ], 200);
     }
+
 }
